Removed ability to use HTML in messages
Before we used innerHTML to render messages. This was useful as people could use tags such as <del>, <b> or <h1> to better render their text. This has a major drawback as it allows them to completly break the page with css and js.
We then changed this to innerText to make sure everything works as expected.
We will probably later make text parsing to allow them to render text in bold or italic.